Module Nbd

module Channel : sig ... end

Channels represent connections between clients and servers.

module Client : sig ... end

An NBD client which allows you to connect to remote disks and perform I/O.

module Mirror : sig ... end

module Mux : sig ... end

Lwt connection multiplexer. Multiplexes between parallel requests from multiple clients over a single output channel to a server that may send responses out of order. Each request and response carries an id that is used to match responses to requests.

module Protocol : sig ... end

Types representing NBD protocol requests and responses.

module S : sig ... end

Common signatures used in the library.

module Server : sig ... end

A Server which allows you to expose an existing block device to remote clients over NBD.