Module Nbd.Mux

Lwt connection multiplexer. Multiplexes between parallel requests from multiple clients over a single output channel to a server that may send responses out of order. Each request and response carries an id that is used to match responses to requests.

module type RPC = sig ... end

module Make (R : RPC) : sig ... end