Module Mux.Make

Parameters

R : RPC

Signature

type client

val rpc : R.request_hdr -> R.request_body -> R.response_body -> client -> (unit, Protocol.Error.t) Result.result Lwt.t

rpc req_hdr req_body response_body client sends a request to the server, and saves the response into response_body. Will block until a response to this request is received from the server.

val create : R.transport -> client Lwt.t

create transport creates a new client that manages parallel requests over the given transport channel. All communication over this channel must go through the returned client.