Module Protocol.Error

Read and write requests can fail with an error response.

type t = [

| `EPERM	(*
Operation not permitted
*)

| `EIO	(*
Input/output error
*)

| `ENOMEM	(*
Cannot allocate memory
*)

| `EINVAL	(*
Invalid argument
*)

| `ENOSPC	(*
No space left on device
*)

| `Unknown of int32

]

Defined error codes which can be returned in response to a request in the data-pushing phase.

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val __t_of_sexp__ : Ppx_sexp_conv_lib.Sexp.t -> t

val to_string : t -> string