Module Protocol.Option

In the 'newstyle' negotiation there is an opportunity for the client to negotiate options with the server. These are the known options.

type t =

| ExportName	(*
The client would like to connect to a given disk/export by name
*)

| Abort	(*
The client would like to quit.
*)

| List	(*
The client would like to receive a list of known disk/exports.
*)

| StartTLS	(*
The client would like to protect the session with TLS.
*)

| Unknown of int32	(*
This option is unknown to this implementation
*)

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val to_string : t -> string