Module Protocol.OptionResponseHeader

The server sends a response to every option request sent by the client (except ExportName which is followed by a DiskInfo. This is the header of the response.

type t = {

request_type : Option.t;	(*
The option type requested
*)

response_type : OptionResponse.t;	(*
The response code
*)

length : int32;	(*
The length of the payload associated with the response
*)

}

The header of the response sent by the server in response to the client requesting an option.

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val sizeof : int

val to_string : t -> string

val unmarshal : Cstruct.t -> ( t, exn ) Stdlib.result

val marshal : Cstruct.t -> t -> unit