Module Protocol.Request

After the negotation phase, clients send I/O requests to the server.

type t = {

ty : Command.t;	(*
The command type
*)

handle : int64;	(*
A unique handle used to match requests with responses.
*)

from : int64;	(*
The start of the data region
*)

len : int32;	(*
The length of the data region
*)

}

An I/O request sent by the client to the server.

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val to_string : t -> string

val sizeof : int

val marshal : Cstruct.t -> t -> unit

val unmarshal : Cstruct.t -> ( t, exn ) Stdlib.result