Module Protocol.Server

In response to a List option, the server sends a number of Server responses and then finally an Ack

type t = {

name : string;	(*
The name of an available disk.
*)

}

A reponse to a List option. Note this option is repeated, once per available disk.

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val sizeof : t -> int

val unmarshal : Cstruct.t -> ( t, exn ) Stdlib.result