Module Nbd.S

Common signatures used in the library.

module type CLIENT = sig ... end

A Client allows you to list the disks available on a server, connect to a specific disk and then issue read and write requests.

module type SERVER = sig ... end

A Server allows you to expose an existing block device to remote clients over NBD.